* uim/uim-scm.c   - (struct cons_args): New type   - (uim_scm_make_int_internal, uim_scm_make_str_internal,     uim_scm_make_symbol_internal, uim_scm_make_ptr_internal,     uim_scm_make_func_ptr_internal, uim_scm_cons_internal,     uim_scm_quote_internal): New static function   - (uim_scm_c_int, uim_scm_c_int_internal): Change return type to long   - (uim_scm_make_int):     * Change integer type to long     * Make callable from non-Scheme context   - (uim_scm_make_str, uim_scm_make_symbol, uim_scm_make_ptr,     uim_scm_make_func_ptr, uim_scm_quote, uim_scm_cons): Make callable from     non-Scheme context

+++ uim/uim-scm.h
@@ -130,10 +130,8 @@ uim_scm_c_bool(uim_lisp val);
 uim_lisp
 uim_scm_make_bool(uim_bool val);
 
-int
-uim_scm_c_int(uim_lisp integer);
-uim_lisp
-uim_scm_make_int(int integer);
+long uim_scm_c_int(uim_lisp integer);
+uim_lisp uim_scm_make_int(long integer);
 
 char *
 uim_scm_c_str(uim_lisp str);
